requirements.txt
discord.py==1.4.1
Pillow==7.2.0
python-youtube==0.6.0
